pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "milady-shrinkifier"
version = "0.2.2"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.121.1",
  "httpx>=0.28.1",
  "numpy>=2.3.3",
  "onnx>=1.19.1",
  "onnxscript>=0.6.2",
  "pillow>=12.0.0",
  "torch>=2.9.0",
  "torchvision>=0.24.0",
  "uvicorn>=0.38.0",
  "wandb>=0.25.1",
]

[project.scripts]
milady = "milady.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["milady"]
